+++ functions/gen-opengraph-image/rollup.config.js
@@ -1,39 +1,31 @@
-import babel from "rollup-plugin-babel";
-import resolve from "@rollup/plugin-node-resolve";
+import babel from "@rollup/plugin-babel";
 import commonjs from "@rollup/plugin-commonjs";
-import builtins from "rollup-plugin-node-builtins";
-import globals from "rollup-plugin-node-globals";
 import replace from "@rollup/plugin-replace";
-import React from "react";
-import ReactDOM from "react-dom";
+import resolve from "@rollup/plugin-node-resolve";
 
 const config = {
   input: "src/image.js",
   output: [
     {
       file: `image.js`,
-      format: "iife"
-    }
+      format: "iife",
+    },
   ],
   plugins: [
     resolve({
-      preferBuiltins: true
+      preferBuiltins: false,
     }),
     babel({
-      exclude: "node_modules/**"
+      babelHelpers: "bundled",
+      exclude: "node_modules/**",
+      presets: ["@babel/preset-env", "@babel/preset-react"],
     }),
-    commonjs({
-      namedExports: {
-        "react-dom": Object.keys(ReactDOM),
-        react: Object.keys(React)
-      }
-    }),
+    commonjs(),
     replace({
-      "process.env.NODE_ENV": JSON.stringify("production")
+      "process.env.NODE_ENV": JSON.stringify("production"),
+      preventAssignment: true,
     }),
-    globals(),
-    builtins()
-  ]
+  ],
 };
 
 export default config;